I15_1-807 (9.10) Fix broken beep on scan end functionality in GDA client
Since this may not be to everyone's liking, also add a toolbar toggle
button to enable and disable the end of scan beep.

Default to off but persist the state in the workspace so it is restored
whenever the GDA client starts.

Rather than showing state using a single icon, switch between distinct
icons for sound enabled and disabled states.

Changes to the JythonTerminalView meant that the old beep on end of scan
functionality became unreachable, so now this has been implemented in a
new SoundToggleHandler, remove the dead code from JythonTerminalView.

Note, since Solstice scans do not generate ScanEvents, only ConcurrentScan
scans are able to beep at scan end.

Change-Id: I250f22347cd5c3cb44032c1d2a0e24edeced5391