DAQ-1990 Experiment plan broadcasts events
Plan already recorded segment transitions & generation of triggers in
its ExperimentRecord instance. Now the beans containing that data are
published over ActiveMQ inside a PlanStatusBean.

PlanStatusBean extends StatusBean and uses the latter's name and status
as name and status of the overall plan. It holds lists of SegmentRecord
and TriggerRecord as well as the name of the last trigger, for conven-
ience. Also stored are driver and driver profile names (if included).
SegmentRecord and TriggerRecord were previously SegmentAccount and Trig-
gerAccount; they are now in the api plugin, their useless interfaces
deleted, and now implement java.io.Serializable.

Tests in PlanTest relating in any way to recording of this data are now
in PlanBroadcastTest; new tests using a mock IPublisher<PlanStatusBean>
are also included.

Conflicts:
	uk.ac.diamond.daq.experiment.test/src/uk/ac/diamond/daq/experiment/plan/PlanTest.java

Change-Id: I92177ebc2e2e8c0fcb8fd06778b2ded0bbb6debc