DAQ-2020 Add messaging API and implementation
This adds packages uk.ac.diamond.daq.api.messaging.* to uk.ac.gda.api
to define an API and some simple messages which can be sent to update
external processes scan progress.

It also adds a new bundle uk.ac.diamond.daq.messaging which provides an
implementation of this interface available as an OSGi service to send
defined messages to ActiveMQ as plain JSON. These messages are intended
for non-Java processes to receive and use. The interface does not
specify the serialization format or the messaging broker to be used so
this could be changed in the future if required.

Change-Id: Ie28ac7292a0daccc1c58d9207367b31de5b0befd