[DAQ-2021] make AcquistionDevice throw InterruptedException if scan aborted
introduce an 'aborted' flag and use this instead of the DeviceState
to check if the scan was aborted. This make the abort mechanism consistent
with the pause mechanism.

We can remove any checks of the deviceState for program control as
AcquisitionDevice is the only class that sets its own deviceState and
we don't set it to any other rest state except at the end of the scan.
(i.e. the end of the run method)

Change-Id: Iae2d8ceeb4c22e5ea3b10c298d6b0241531bec6d