DAQ-2026 Add PlanProgressPlotView to plot progress of experiment plan
This view subscribes to EXPERIMENT_PLAN_TOPIC messages and shows the
progress of a running plan. The main component is a plotting system
which shows:
 - the name of the plan and its status (RUNNING/COMPLETE) as plot title
 - the driver profile (i.e. the demand)
 - the actual trajectory, sampled at 2 Hz
 - segment transitions as vertical markers
 - triggered events, organised by triggers
 
There is also a text box showing the plan name, its start time (its
completion time when it's finished), the current segment name and its
start time.

Still to do is to mark the success of trigger events, which is already
broadcasted.

Change-Id: Ica51a75e2e95ba7ff1e665ae7b088c79b29cb81e