[DAQ-1840] ColumnDataFile: ensure "" is treated as Dimensionless
To specify a conversion to/from a Dimensionless quantity, you have
to write two double-quotes in the input file. The ColumnDataFile
class passes these through to the Unit class.

When using JScience2, Unit seems to be able to parse this string,
but it fails with JScience4 (which is probably the correct behaviour).

This changes the two double-quotes to an empty string, to make the
conversion explicit.

Change-Id: I26d7d60c1d009dab90ab12f128d580ae1c3cf43e