DAQ-1995 - Dynamic RMI don't warn when using the Finder on the server
This is needed to partially address DAQ-2055 however the behaviour
should be changed anyway because its might be the case you want to see
if an object can be found, and currently if you do this a spurious
warning will be issued.

Also downgrades some verbose logging messages to debug.

Change-Id: Ia7b369a4ca9d5a0733586921208f828f482292a2