[DAQ-2056] MalcolmDevice.validate should work outside of a scan
There in an important distinction between an IMalcolmModel which is the
model in GDA for a MalcolmDevice, and the model that an actual malcolm
device expects in a call to configure or validate. As malcolm performs
the (inner) scan, it needs to know some things about the scan,
specifically the point generator and the output directory to write to.

Change-Id: Ic3789371e92df097086f2e100a6dd5f5d91b0d7b