DAQ-2025 Make updatePlotRegionFrom method run in UI thread
The addRegion and setRoi calls in the updatePlotRegionFrom method throw
Exceptions if not run in the UI thread, which requires client of this
method to make the switch. This usually happens distant from this method
and via various observable/event calls which can thus lead to races
between the Plot update happening and the calling method proceeding.

This change puts the thread switch as close as possible to the methods
that require which prevents this.

Change-Id: If8a8dd9ec794f78fb7aa779a6d36d0ca39a289d8