MXGDA-3218 Simplify logic for assigning baton (or not) to first user
Set firstClientTakesBaton = true if !useBaton. If we're not using baton
then this client is the only client and apparently in this situation we
give the "singleton" client the baton.

Doing this allows us to fix the if/else, which executes the same code in
both branches.

A UDC client will take out a lease but it can't take the baton, so when
a "human" client comes along there's already a leaseholder so they won't
get the baton at statup. Instead of checking if there are any
leaseholders, check isBatonHeld(), then human user will get baton.

Change-Id: Ib96f2fb9548fd850c6559e13211ee783201eb41c