[DAQ-2057] add test for malcolm valdiation where the model is invalid
and assert the status and message of the bean is as expected.

A previous change made it the case that ScanProcess.execute does not
throw an exception in most cases, e.g. if the scan fails or if validation
fails. Instead it sets the status of the ScanBean to FAILED and updates
the error message.

Change-Id: I565a475fd08e3f2f6f90357383a5a56f6941d51e