[I14-618] Improve XANES scripting interface (client side)
The purpose of this change is to create a more more robust
interface to allow user scripts to control  XANES Mapping scan
submission. The same technique can be used in future for
tomography & ptychography scans.

In contrast to the previous version, this passes ScanRequest
and XanesEdgeParameters objects to the user script to manipulate,
rather than an mscan command that must be parsed and reconstructed.

- Move XanesEdgeParameters to API bundle so it is accessible
to the server
- Factor out ScanRequest generation so XANES submit button can use it
- Use ScriptService to send ScanRequest & XanesEdgeParameters objects
to user script rather than mscan command

Change-Id: Ia141cc037db36dee07c4339809aa2cead049ba5d