Add NexusScanFileHandler and DatFileHandler
Now that the DetectorFileDisplayer uses composable file handlers,
this offers a handler to display scans created using the standard
ConcurrentScan in GDA. The two axes marked as @primary and @signal
are used as x-axis and y-axis respectively.

For dat files, the first column is assumed to be the x-axis and all
other columns are plotted as y-axes.

Change-Id: Iffba7f969ee73b39790483238aa92967f38eb29f