[DAQ-1465] add JmsQueue reader to read JMS queue and submit to consumer
Note: this is in preparation for bringing the queue into memory. It
cannot be used yet as it would in effect be submitting the bean to the
same JMS queue again. Once the queue has been brought into memory
this class allows all the JMS logic to be removed from the consumer.

Change-Id: I389b8591d0d34cb04a97ec4bf3f6175bf19953e1