DAQ-2126 : Prevent Fluorescence detector view collections from interfering with scans
- A server side object has been created that returns MCA data for a fluorescence
detector only if scan or script is not running. This object is exported to
the client and used to retrieve the MCA data. This should prevent the
problem of the client being informed of scan starting too late, and
continuing to collect data and interfering with the scan collection.
- The FluorescenceDetector view Live mode will now not continue to run if an
exception is thrown during the collection.
- Updates to Xspress (buffered) detectors so that they wait at beginning
of scans (before clearing scaler memory, frames etc) for the getMCAData
function to complete.

Change-Id: I5939e3033baba13b742f6d536d26ac0a7fba9068