B18-369 : Allow DetectorRate view format to be customised
This change allows the number format of the fields in the detector
rate view to be customised, and allows them to have a different format
from that displayed during scans.
- A 'numberFormatMap' has been added to client-side DetectorRateViewConfig.
This stores the output format to be used for quantities in 'extra names'.
(i.e. key=field name in 'extra name' of scannable, value=format)
- When creating list of string values to be displayed, the numberFormatMap
is used to replace parts of a scannable's outputFormat if a format for
particular 'extra name' quantity has been set.

Change-Id: I8feaf50d006cd3688f5f9c3e413793e94685c7e8