[DAQ-1840] Convert to use JScience4
General changes:
- required bundle in launchers, manifests & features
- put JScience4 above january in Ant build files classpath, to ensure
that the JS4 versions of packages are used rather than the version
exported by January
- package names for quantities & units
- Amount<? extends Quantity> instead of Quantity
- getEstimatedValue() instead of getAmount()
- define ZERO_LENGTH etc. as the quantity classes no longer do so
- implement alternative strings for creating units
(e.g. "deg", "degrees") as aliases not alternates.
Alternates cannot be created for non-SI quantities in JS4.
Note that aliasing means that a Unit does not remember the string
from which it was created, so it will always return the "standard"
string. A number of tests have had to be changed to reflect this.

Changes specific to this repo:
- Reinstate Serializer tests that failed with JS2. Ignore the decibel
test which now fails in JS4, but is not relevant to GDA
- Rename some functions in GDA-defined quantities (e.g. BraggAngle),
as type erasure means that they must have distinct names

Change-Id: I4f39a913477cc0a4f81b9e6257fad0cf58265260