B18-377 : Spreadsheet view : allow detector configuration file to be changed
This change allows the spreadsheet view to control the name of the detector
configuration xml to be used for Detector parameters. File name selection
is via a combo box listing the xspress2, xspress3 or xmap files in the
currently selected base xml directory.
Other changes :
- Content of each file name combo box is now updated dynamically each time
it is opened using latest list of xml files on disk.
- The column widths in the table are now adjusted automatically to fit
the contents when the first scan is added.

Change-Id: Ie7b9a33f5fc9dee88617580b67beb7ccd0da7ff6