[I12-562] Add DummyDXPModel
This is a very basic dummy version, returning fixed values.

A better solution would probably be to extend the server-side
XMAP object (in both live and dummy versions) to handle acquire status
and dead/live/real time and to use this via RMI in the client.
This change is out of scope of the current task.

Change-Id: I79ba2b7b2d97f64eab99498ba98480f2c0abb94c