Make Jython Findables module writing configurable
Make the Jython Findables module generation and writing configurable via
three GDA properties:

  gda.jython.findables.module.enabled
    Whether to generate and write the module at all (default: true).

  gda.jython.findables.module.name
    The name of the module (default: gdaserver).

  gda.jython.findables.module.dir
    The directory where the module should be written (default:
    ${gda.config}/scripts).

This provides a way to turn off the generation of the module altogether,
change the name of the module, and change where the module is written.

The defaults are such that it behaves just like it did before this
change (i.e., it's a backward-compatible change).

Change-Id: I91917cf49da0237694179668bf30b80925a5f21b