DAQ-2024 Add support for descriptive filenames
Add a filename factory and browser composite to allow the user to see
the shape and path of the scan saved within a .map file without the need
to open the file. A descriptive section is automatically added between
the name and extension for this purpose. The browser support both this
format and files saved as <name>.map.

This will allow the user to easily select and load previously defined
scans by simply double clicking them in the browser.

Change-Id: I6e696b3dd9d8b0d6ff008e99afa00a21207ff9d9