DAQ-2142 Call Malcolm.reset() at the end of a Malcolm scan
A recent Malcolm refactor for reliable top-up pausing requires a
reset() call at the end of the scan to actually close the file.

MalcolmDevice.reset() is now called when invoking @ScanFinally.

Incidentally, reset() as called during configure() used to call abort(),
then reset() with some custom exception handling. This isn't the desired
behaviour at the end of the scan, so reset() is here changed to do only
that, and the call in configure() is replaced with goToReadyState(),
which does that.

Change-Id: I9efb4d31a8aa2d2aa67c7e4e4d2ae1631033cb6a