Move SSH connection handling methods out of JlineSshServer
This creates standalone implementations of command handlers so that the
server can just handle the creation of connections and not starting shells.

JlineSshServer no longer had any dependence on Jline so has been renamed
GdaSshServer.

Some of the functionality of JLine's command implementation has been
duplicated allowing us to better control the threads used and the
logging around SSH connections.

Change-Id: I2d3da274fa0097f9787528d8aad2d137ad8b4dff