[DAQ-1704] rework queue to use two maps, in preparation for using persistence
The library that we're going to use to reintroduce persistence
MVStore uses a map type, MVMap which works like SortedMap in that it
keys can be navigated by natural order, but not by insertion order
unlike LinkedHashMap. The solution is to use an additional map
to hold the order of the queue as a map from queue position to bean id.

This change exits is split out the use of two maps and making sure the
logic is correct before adding persistence back in with MVStore.

Change-Id: I564fe973e46e9a743c93e94ca0b700b556e65e28