DAQ-2027 Link Trigger to Scan
Before this change, the only way to determine which scan results from a
particular trigger is by finding a pattern with timestamps.

In this change, Triggerable.trigger() is changed to return Object (was
void). The concrete Java implementation for Solstice scans (Triggerable-
Scan) will then return a ScanBean whose unique ID is broadcast in the
TriggerEvent. 

A view to monitor the progress of the plan can then subscribe to both
the plan and the status topics, and link triggers from the first to
scans from the second using the same unique ID.

Change-Id: I120a643cd8c79324312aca4270239e296ade2979