DAQ-2164 (9.13pre) Ensure submitQueue is immutable by removing setSubmitQueueName
EventServiceImpl was changed to use queueName as a key into a Map of
consumers, to prevent multiple instances of the same consumer being
created. Unfortunately the submit queue was also mutable, which allowed
a consumer to be switched to a different submit queue after creation.

This combination allowed a default queue to be created and modified,
preventing the default queue from being recreated elsewhere.

This change forces the submit queue to be set in the constructor and
nowhere else, so the map can never get out of synch with the consumers
it contains.

Change-Id: I1b018a80046562942ceca28a4ddce1bad23b9189