DAQ-2088 Delete junit suites in scanning tests
These were once used to determine which tests should run and as ways of
grouping tests into categories to run at the same time.

At the moment the tests that run are defined in releng.ant.
It is also easy in Eclipse to run all tests within a project/package so
the Suites and 'All' are not necessary.

Change-Id: Icbf8619b344d6923a3cda11a9d1d7c2c9e1e27f2