[DAQ-2181] fix acquire tests after january change
An change in the updated January library broke how SolsticeScanMonitor
wrote the pointStart and pointEnd datasets in the case of an acquire
scan (where the scan rank is 0). Using intializeLazyDataset, which takes
the scan rank (which is always a minimum of 1) seems to fix this.

Using initializeLazyDataset means that the datasets shape is not fixed
at the start of the scan, but expands as required. This required a fix
for SolsticeScanMonitorTest.

Change-Id: Ifb5cab6edbf16364e2cd90e1338e964980234a82