DAQ-2213 - Allow alternative unicode code points for "micro" unit prefix
The "lower case mu" and "micro" symbol are different Unicode code points,
even if they effectively render the same.

micro is U+00B5
mu is U+03BC

We should aim to support both code points for the micro- prefix, as
as well as the two code points for the Angstrom symbol.

The names for the String variables in NonSIext.java are adjusted as
well, for the sake of consistency, and non-ASCII characters are
replaced in variable names.

Change-Id: I8d7dcb3972179f50e6639fb8aa7b79721ed847f4