DAQ-2202 Dark/flat field collector for mapping scans
New class CalibrationFrameCollector does the following when included in
a mapping scan:
 1) configures the beamline to achieve desired dark/flat/whatever
	conditions
 2) acquires a single frame with the detector used for the scan
 3) restores the beamline to effectively undo step 1

The frame is then appended to the NeXus tree as an NXdetector under
/entry/instrument/nameOfCollector/nexusFieldName.

In the case of Malcolm scans we currently don't know what the actual
detector used is and we need this to retrieve the frame. The scannable
therefore can be configured with a map of malcolm name (i.e. MRI) to
detector name (e.g. andor_addetector)

Also in this change, a new mapping view section configured with dark
and/or flat collector names to include/exclude from the scan being
defined.

Change-Id: Ib7ee3824eeae767b52fc9b3c99828d5eea589803