I07-169 Only use previous settings when Y axis configuration is the same
When gda.gui.scanplot.maintainYTraceSelectionOnXaxisChange is set
we want to use the previous line setting but not if the YAxesShown/NotShown
is different.

This used to be handled by scanPlotSettings#equals however this check
is skipped when gda.gui.scanplot.maintainYTraceSelectionOnXaxisChange is
set to true.

Also add Javadoc to usePreviousScanLineSetting.

Change-Id: I68dd9c41ff617adfceb24f332684b056c9a20213