Bug fix: Don't call reset() at the end of a scan with Malcolm 3
DeviceState.FINISHED is the state introduced in Malcolm 4. Earlier ver-
sions will fail when GDA calls reset() at the end of the scan. We can
therefore check the device state and only call reset() if needed. This
is not a race condition because if device state is FINISHED it will stay
until reset() is called.

Change-Id: I3fad9b01da07a1974785c035ac42737cbeb2b3c1