DAQ-2230 Ensure EDXDController only configures itself once
Also move status channel update handler into a separate function

Creating a scannable and adding it to a ScannableGroup (as is common on beamlines)
results in the scannable's configure() function being called twice, as
ScannableGroup calls configure() on all its members.

This can cause a significant problem, as configure() adds all the
subdetectors to a list, which can lead to duplication.

Change-Id: I30e39aa02e661a6f683cf8ef4c4d48a7746f146e