I16-351 (9.13) Suggest solution when Camera configuration can't be found
The LiveStreamView relies on CameraConfiguration beans being findable,
but the LiveStreamMenuContribution doesn't, so the latter will happily
add menu contributions which fail with a "Camera configuration could not
be found for camera ID null" error when they are used, if the client
spring doesn't contain a FindableNameSetterPostProcessor bean.

Detail this in a warning so that this problem can be fixed more quickly.

Change-Id: I0f7c6712233e957844b8e8c688932a1f242e1da7