I20-441 : Allow I0 channel used by TfgFFoverI0 to be determined automatically
The I0 data channel used by TfgFFoverI0 when computing FF/I0 can now be
determined automatically. This can be useful when the output fields of the
counterTimer are changed from scan to scan and the I0 field might not always
be in the same position.
- The counterTimer extraNames are examined and the index of the data channel
is assumed to be the first one not called 'time'
- This behaviour can be activated by setting the I0_channel to -1.

Change-Id: Idc0ba37ecaf8225f96b409c269b3cdb783a4e073