DAQ-2191 Initial file system based implementation of experiment service
Supports saving/loading experiment plans and driver profiles. Extends
DummyExperimentService until scan support is added.

Objects are serialised into JSON and saved in gda_var/EXPERIMENT_BASE/

... note that experiment ID is currently not used.

An Activator was added to the bundle to access IMarshallerService and
IFilePathService.

This also exposed some GUI bugs in ProfileManagerPage and SegmentEditor;
both are addressed here but the latter will need a better solution (it
is marked with a FIXME tag).

Change-Id: I4974b5551c58baf174ac8f88b6151d2f1b08cc13