DAQ-2191 Add view to manage experiment plans
This change adds a new view allowing users to create, edit, delete and
run experiment plans.

The view consists of a list viewer (actually a tree viewer pretending
to be a list viewer) which shows the available plans. Selecting one
enables the Run button.

Below the list there are buttons to add, edit and delete. The first two
open the plan setup wizard, either with a new plan, or the selected one.
The delete button requests user confirmation before deleting the plan.

Icons used before in this plugin have been added to ExperimentUiUtils.

Change-Id: Ieae8b8aa1c9af35b1806dbc7561b0adf6acef523