I16-395 (9.13) Add property to disable pos defaulting to all scannables
On some beamlines, the default behaviour of pos when no arguments are
provided is sub-optimal. Iterating over every scannable can take a long
time, especially if scannabes hang on readback. In this case, stopping
each hung scannable individually can be frustrating.

This mechanism is intended to be a more robust replacement for the pos
function in I16's localstation.py which has recently become unreliable.

Change-Id: I8fcb4b9f352f9c2962252277f246fff4f4d48c85