Fix product builds for non-mapping beamlines.
As a result of the new nexutemplate support only being included via the
mapping ui feature, non-mapping beamline produc sites cannot build. This
change adds the plugin to the client feature, which all beamline clients
use, to allow the builds to succeed. The correct place overall will be
resolved when Matt Dickie is back from holiday.

Change-Id: I30e95e9b078040f5ea2c3234d81f18af02d8affc