B16-185 Add Camera control for live stream
Detectors which sometimes are used in scans (where e.g. they are
using single image mode) also want to be used with the live stream view.
This requires continuous image mode and an acquisition period to be set
which is in keeping with the exposure time.

This class extends EpicsCameraControl and overrides two of the methods
to set the required state of the detector.

Change-Id: I17df339365e532e44e474be2f8cf8c6345ef5159