DAQ-2129 View which plots detector's latest frame for the running scan
A very simple view consisting of a single plotting system. But how do we
get the most recent frame?

The LiveFileService (used by DataVis) provides its listeners with the
LoadedFile corresponding to the ongoing acquisition, and requests that
it be refreshed whenever new data is available.

This change adds a new listener implementation: LatestSwmrFrameFinder.
The listener traverses the uniquekeys dataset to find the position of
the latest data point, then slices the detector data accordingly.
This frame is given to some Consumer, which in the case of the view
is a plotting function. The refresh rate is limited to 0.5 Hz which is
what the Mapping/FileController/DataVis visualisations use.

Change-Id: I1d9214c40fd9d44d7feac66fd3c67f26e6357837