I20-450 : Make 'extra columns' of data work for Xes scans
The 'signal list' objects representing the 'extra columns of data' are now
added to the scan parameters in XasScanBase.createAndRunScan using the
output preparer rather than in 'createScanArguments' method of the different
scan types. This ensures the signal parameters are always added if needed
whatever the scan type, and are only added once.

Change-Id: Ia9b7e1cf892f27e5148a5108c6c10a2641b27c66