DAQ-2318 Ensure consistent driver/profile selection
This change extracts the experiment driver and its profile fields from
the ExperimentPlanBean into a separate class so these may be set at once
for data consistency.

To keep the wizard page responsive in terms of enabling/disabling navi-
gation buttons, MetadataAndExperimentDriverPage now listens to its UI
components, rather than the bean (which, as a consequence of the change
described above, doesn't always change whenever users interact with the
UI).

Change-Id: I4d6ab98000b4101dd6cdea9158fc56a8de105e73