Remove SoundToggleHandler::dispose, preventing an extraneous JSF
On client shutdown, the JythonServerFacade is disconnected before
handlers have their dispose methods called. Calling "getInstance"
after this will result in the creation of a "new" facade, on which it
doesn't make sense to unregister the old observer.

As a side-effect (in conjuction with changes for Unattended Data
Collection), creating a new JSF will result in the currently
exiting client being reassigned the baton, creating a sort of "ghost
client" that cannot relinquish the baton (it eventually times out,
allowing other clients to take it).
This is likely the cause of the numerous baton issues reported by
MX beamlines (e.g. MXGDA-3336).

Change-Id: I950445af5453522c3ef4b06750f3b3c61b73fcc5