DAQ-2280 Add support for 2D line and point scans to new mscan
Adds the necessary models and regions to AreaScanpath and RegionShape
and update associated tests. Update the MScanSubmitter and test and the
mscanHandler script. Also update the validateAndAdjust method in
ClauseContext and add tests.

The majority of the other changes are associated with the renaming of
the old Roi class to RegionShape to reduce ambiguity, and all the
associated refactoring.

This is exclusively the case for:

IClauseElementProcessor
ReadoutCheckerTest
RegionShapeElementProcessor
RegionShapeElementProcessorTest
ResolutionTestsBase
Roi
RoiElementProcessor
RoiElementProcessorTest
RoiTest
ScanClausesResolverTest
ScannableElementProcessorAndElementProcessorBaseTest
ScannableGroupElementProcessorTest

Change-Id: Ib51be69e46d0bc1781bc2a4e247614b4fad04c77