Remove warning from NDW when DataWriterExtender is added
For beamlines that want to write both Nexus and SRS files the createSrsFile
is true. This warning is logged during every scan but is most likely
a non issue.

Delete whole method because it would only be calling super implementation
now.

This was originally added in 41c65afb15 as part of GDA-5852 to warn about
a potentially breaking change.

Change-Id: I76c22089ad4e19511d29bc9743d916a6d476ce28