DAQ-2368 Remove logging regarding SRS file from NexusDataWriter
This message was placed before the SRS file's url was set hence the log
message was always:
    Also creating file (txt): null

Since the SRSWriter logs to the terminal when it is writing a file
this should be sufficient. The intention was to log using the logger as
well so add this to IncrementalFile.

Also some minor logging/squid tidy ups in IncrementalFile.

Change-Id: I059c9ab52387533cd01139c181359bc006f8bb3f