DAQ-2396 Fix acquire requests for Malcolm scans
As of Malcolm 4 a static scan containing no axis is a valid SPG. There
were a couple of issues in AcquireRequestHandler which would prevent us
from doing such scans.

When the detector is an IMalcolmDevice, the static generator needs to be
wrapped inside a compound generator. This would not work for a GDA scan.

Additionally, we need to pass the scan model when invoking PreConfigure
so that the filepath and model can be configured

Change-Id: If223325aa94c0dce1541b60a1c72769482b61553