I07-247 Don't fail scan if a metadata scannable has been deleted
If a scannable is deleted from Jython namespace it still exists in the
set of scannable names to collect metadata for. This is obtained from a
NXMetaDataProvider however the method didn't check that the object
existed so could end up with null in the returned Vector.

This method is probably only called by beamlines that have the
standard MetadataCollector as a default scannable in order to
write consistent metadata to the SRS dat file header and the Nexus
file.

The Nexus writing of metadata already checks for this in
NexusDataWriter::makeMetadataScannables.

Change-Id: I81f1abc7583c9f3cb7a1fced28d30bd840103e66