I20-457 : Improve layout of 'Fluorescence detector' view element table
The total height of the detector element table in the 'Fluorescence detector'
view is now set based on the actual height of a row. Previously, a
fixed height was used which was sometimes smaller than the actual height
which would result in 'missing' table rows.
The width of each row element has also been slightly increased, so that
the text appears with correct horizontal centring.

Change-Id: Ibfc9e3a401e7e2a31706583c575826e81de95c23