DAQ-2451 Trim whitespace from metadata scannables
Scannables that have the output format specified (e.g. %5.5g) causes
data to be aligned properly in the Jython console when a scan is in
progress. However in the dat file this leads to entries such as:
  dets3top=-1.4532
  d4x=  -66
  hex1y=    0

Remove this by trimming the position value string.

This object is used to insert metadata into an object in the Jython
namespace which is picked up by the dat file writer.

Change-Id: Ib6881f76af9e84170e8b0c277d9ca95f0ac2b468