DAQ-2422 Support PyString in NXMetaDataProvider
When a Jython scannable returns a string from a getPosition call it is
already available a Java String. This handles the case when a scannable
has multiple components/extra names. In this case the result from
getPosition is e.g. a PyList of PyStrings.

Also remove a check for PyList since it extends PySequence.

Change-Id: Ic1329d4aaa81ce8353d373c450ed06cd4aaefd3e