Removes incorrect jre.compilation.profile from bundles
jre.compilation.profile was set to Java 1.6 or Java 1.7 in several
bundles. This will fails any Tycho build as the
Bundle-RequiredExecutionEnvironment (BREE) in the MANIFEST.MF of the
affected bundles are already configured to use Java 8 and Tycho prefers
jre.compilation.profile over the BREE.

By removing it, we ensure that the BREE from the manifest is used by
Tycho.

Change-Id: I1da41b701d5adc3f56d8d67f0c445cf619c1a011
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>