Add ScanInformation to scan start calls for NcdSubDetectors
Several implementations require information about the current scan and
currently get it individually. This gets it in one place to prevent the
need for extra calls and to ensure all sub detectors use the same
information.

The new method is added with a default so implementations that do not
require it do not need to be changed.

Change-Id: I45942358a4c4417026ba66ab77aa15e040193fb2