DAQ-2475 Fix "continuous" checkbox binding
The previous behaviour meant that when the control should be disabled
(because we have a software detector selected) the field in the model
is also set to false and vice versa. Consequently, loading a Malcolm
step (i.e. non continuous) scan from the queue would:
 1) enable the checkbox
 2) set the selection to true!

The previous behaviour was logical in that when the control is disabled
it should also be unchecked. This change preserves that visual behaviour
without modifying the model so that when the control is enabled, the
previous, deliberately-set state is restored.

Change-Id: Iacaca98336ac3f634673d307c5e9590214e312d8