DAQ-2031 : Generic detector preparer
- A new DetectorConfig bean has been created. This holds settings for
configuration of a detector: name of the detector object,
whether to use it in a scan, XML configuration file to use,
Jython command that should be run before configuring the detector.
- Extra functionality has been added to DetectorPreparerFunctions
to configure and prepare a detector for a scan using settings from
a DetectorConfig object. Detector configuration (e.g. to set ROIs, scaler/MCA
options etc.) is done using same XML files as are currently used.

Change-Id: I50bf88a84904a9604864800603817620f87812b6