DAQ-2500 Initialise NumberAndUnitsComposite with scannable units
Until now, this control has been initialised with hard-coded units for
length (regardless of the scannable describing the quantity). There are
two mechanisms for overriding initial units. The first is a property
which overrides units for all fields (start, step, etc). The second is
InitialLengthUnits which allows configuring units per field (e.g. start
in mm, step in um).

With this change, we ask the scannable for its units, and populate the
compatible units given the dimension of the quantity. However it does
create a problem with the aforementioned mechanism for overriding units
since they assume a length dimension. For now, the solution is to only
go through those mechanism when the scannable in question describes a
length. Additionally, where the default value of these mechanisms was
hard-coded as "mm", it is now injected by the caller with the scannable
unit.

All this units stuff is taken out of AbstractModelEditor into new class
UnitsProvider in the same package.

The compatible units in NumberAndUnitsComposite are now given as a Set
instead of a List.

Change-Id: I8078b4b849e47f579d8a825490cab9a4c8903279