DAQ-2460 Convert between rectangular mapping regions if necessary
When a scan is defined in the Mapping GUI as a CentredRectangle, it
will be converted to the equivalent Rectangle object when it is
submitted to the queue.

When opening such a scan from the queue, GDA used to fail for beamlines
that defined an editor only for CentredRectangle, not for Rectangle:
currently, these are K11/DIAD and I14.
The failure could also happen when re-opening a client without resetting it.

This modifies RegionAndPathController so that, if you try to set
a Rectangle scan shape and only a CentredRectangle editor is available,
the Rectangle will be converted to the equivalent CentredRectangle.
It also handles the opposite case (CentredRectangle loaded when only
Rectangle editor available), though this case is unlikely to occur.

This change is an interim fix, pending a more thorough redesign of
the handling of MappingBean and ScanRequest.

Change-Id: I22638b8ca4330009ccc45f07fb2b4b2cb7572753