Exit progress sub-shell if parent shell has exited
Progress output is executed in a sub-shell in gda_servers_core, but
if the parent shell exits (e.g., due to receiving a signal such as
SIGTERM), the progress sub-shell will continue running indefinitely if
it has not detected that the status port is open.  Fix this by checking
in the progress sub-shell that the parent shell still exists, and if it
does not, exit the sub-shell.

Change-Id: Ie917c4de4bfc7be2095cf7c97e53d248bfab050e