DAQ-2521 NDPythonImpl: do not set file/class names at startup
Revert most of 770fef4cd94faf22c1aacff90151df1d74747319
as this caused GDA server startup to fail if file and/or class names
could not be set (e.g. if the IOC was not running)

File & class names cannot therefore be set in Spring - but this was
only done on I14 and has now been removed.

Change-Id: Ied44279a8c5adc23878f23512c6a4244684a3c30