DAQ-313 Implement map view buttons as toggle buttons
This is not entirely straightforward as:

(1) The initial state of each button should ideally be set from the
state of the system in PlottingController & BeamPositionPlotter, but
this is not available to the Eclipse framework that creates the
respective command. It is therefore hard-coded to true, because the
initial states currently default to true. If the defaults are
changed in the Java, the state in plugin.xml must be changed accordingly.

(2) Some operations change other settings as side effects.
For example, making the mapping path visible also makes the mapping
region visible, which can result in the state of the buttons/menu items
getting out of sync with the plot. To try to minimise this, instead of
toggling the state of the Eclipse command when it is executed, the
state is read from the relevant controller, to bring the two into step.

Change-Id: If83f6ea0a484e8c3154ad3ecf9019acbc6f4859b