DAQ-2432 Run processing based on dataset sizes from acquisition not scan info
The dimension sizes from the scan information itself will not match the actual
data acquired if the scan has been fast forwarded.

We would like to run the processing if a scan has been fast forwarded.

Move the scan dimension check (which does still require the scan information)
to the start of the method so that datasets are not needlessly created if the
processing is not going to run anyway.

Change-Id: If27fedac289c14a810d17e9b0dec492eb58fe7d2