DAQ-2565 Switch between raw/mpeg streams in SimpleLive Stream
Most of the changes done are to implement an element
in the CameraConfiguration capable to switch between
multiple cameras and streaming type, as well start/stop listening
at a live stream.

Some help for the reviewers: The starting point are
CameraConfigurationDialog and CameraConfigurationDialog.
These classes assemble the GUI and contains a Composite
created by the the StreamControlCompositeFactory class.
This class is associated with a State pattern (IdleState, ListeningState)
These states publish Stream events (ListenToConnectionEvent, StopListeningToConnectionEvent)
on which is listening the LivePlottingComposite
(the live view on the left side of the CameraComposite), which in cascade publish a
PlottingSystemUpdateEvent on which is listening the HistogramComposite.

All the events are managed by Spring: any instances can publish or add a listener
using SpringApplicationContextProxy

The LiveStreamConnectionManager is a singleton which keep track of all the existing
connections

Multiple cameras can be configured in a property file using a dot-formatted string
(see K11-config client/_common/common_instance_java.properties)

Change-Id: Ia24852e1f6dc9849af1b6cb72713b1e8171693f7