I14-758 Better handling of non-XRF processing files in XANES GUI
The code extracts "lines to track" from the selected processing files.
However, there may be files selected (e.g. for Merlin) that do not
contain XRF line data.

This change avoids a NPE if there is no XRF data, and downgrades
the error message to a warning, as selecting non-XRF files may
be intentional.

Change-Id: I9fdfa268a9d2a495716b2d8fc4daebb58a7c1b9b