Fix NPE when attempting to release IncrementalFile
If a scan is aborted/shutdown early in the first point (e.g. if a
detector's prepareForCollection throws) then the file is never created.
This only happens when NDW is used in addition to SRS files because in
SDPP a different method, completeCollection, is called. In NDW it
doesn't call this on the SRSDataFile so perhaps that should be changed
instead.

Change-Id: I3f6d12004f62539ae48c9e4dfccf0258a17c4d69